Rename InitialDataTci.?pp to SetInitialRdmpData.?pp
Also renamed tests
