+++ src/Evolution/Executables/Burgers/EvolveBurgers.hpp
@@ -51,8 +51,8 @@
 #include "Evolution/Systems/Burgers/FiniteDifference/RegisterDerivedWithCharm.hpp"
 #include "Evolution/Systems/Burgers/FiniteDifference/Tags.hpp"
 #include "Evolution/Systems/Burgers/Subcell/GhostData.hpp"
-#include "Evolution/Systems/Burgers/Subcell/InitialDataTci.hpp"
 #include "Evolution/Systems/Burgers/Subcell/NeighborPackagedData.hpp"
+#include "Evolution/Systems/Burgers/Subcell/SetInitialRdmpData.hpp"
 #include "Evolution/Systems/Burgers/Subcell/TciOnDgGrid.hpp"
 #include "Evolution/Systems/Burgers/Subcell/TciOnFdGrid.hpp"
 #include "Evolution/Systems/Burgers/Subcell/TimeDerivative.hpp"
