+++ src/Evolution/Executables/ScalarAdvection/EvolveScalarAdvection.hpp
@@ -52,8 +52,8 @@
 #include "Evolution/Systems/ScalarAdvection/FiniteDifference/RegisterDerivedWithCharm.hpp"
 #include "Evolution/Systems/ScalarAdvection/FiniteDifference/Tags.hpp"
 #include "Evolution/Systems/ScalarAdvection/Subcell/GhostData.hpp"
-#include "Evolution/Systems/ScalarAdvection/Subcell/InitialDataTci.hpp"
 #include "Evolution/Systems/ScalarAdvection/Subcell/NeighborPackagedData.hpp"
+#include "Evolution/Systems/ScalarAdvection/Subcell/SetInitialRdmpData.hpp"
 #include "Evolution/Systems/ScalarAdvection/Subcell/TciOnDgGrid.hpp"
 #include "Evolution/Systems/ScalarAdvection/Subcell/TciOnFdGrid.hpp"
 #include "Evolution/Systems/ScalarAdvection/Subcell/TciOptions.hpp"
