+++ src/Evolution/Systems/NewtonianEuler/Subcell/SetInitialRdmpData.cpp
@@ -1,7 +1,7 @@
 // Distributed under the MIT License.
 // See LICENSE.txt for details.
 
-#include "Evolution/Systems/NewtonianEuler/Subcell/InitialDataTci.hpp"
+#include "Evolution/Systems/NewtonianEuler/Subcell/SetInitialRdmpData.hpp"
 
 #include <cstddef>
