+++ src/Evolution/Systems/ScalarAdvection/Subcell/SetInitialRdmpData.cpp
@@ -1,24 +1,25 @@
 // Distributed under the MIT License.
 // See LICENSE.txt for details.
 
-#include "Evolution/Systems/Burgers/Subcell/InitialDataTci.hpp"
+#include "Evolution/Systems/ScalarAdvection/Subcell/SetInitialRdmpData.hpp"
+
+#include <cstddef>
 
 #include "DataStructures/DataVector.hpp"
 #include "DataStructures/Tensor/Tensor.hpp"
 #include "DataStructures/Variables.hpp"
-#include "Evolution/DgSubcell/ActiveGrid.hpp"
 #include "Evolution/DgSubcell/PerssonTci.hpp"
 #include "Evolution/DgSubcell/Projection.hpp"
 #include "Evolution/DgSubcell/TwoMeshRdmpTci.hpp"
-#include "NumericalAlgorithms/Spectral/Mesh.hpp"
-#include "Utilities/Gsl.hpp"
+#include "Utilities/GenerateInstantiations.hpp"
 
-namespace Burgers::subcell {
-void SetInitialRdmpData::apply(
+namespace ScalarAdvection::subcell {
+template <size_t Dim>
+void SetInitialRdmpData<Dim>::apply(
     const gsl::not_null<evolution::dg::subcell::RdmpTciData*> rdmp_tci_data,
     const Scalar<DataVector>& u,
     const evolution::dg::subcell::ActiveGrid active_grid,
-    const Mesh<1>& dg_mesh, const Mesh<1>& subcell_mesh) {
+    const Mesh<Dim>& dg_mesh, const Mesh<Dim>& subcell_mesh) {
   if (active_grid == evolution::dg::subcell::ActiveGrid::Subcell) {
     *rdmp_tci_data = {{max(get(u))}, {min(get(u))}};
   } else {
@@ -31,4 +32,14 @@ void SetInitialRdmpData::apply(
                       {min(min(get(u)), min(subcell_u))}};
   }
 }
-}  // namespace Burgers::subcell
+
+#define DIM(data) BOOST_PP_TUPLE_ELEM(0, data)
+
+#define INSTANTIATION(r, data) template struct SetInitialRdmpData<DIM(data)>;
+
+GENERATE_INSTANTIATIONS(INSTANTIATION, (1, 2))
+
+#undef INSTANTIATION
+
+#undef DIM
+}  // namespace ScalarAdvection::subcell
